refactor: remove SysProcAttr and terminate child processes from parent when shutdown
