+++ cmd/up/exec.go
@@ -22,6 +22,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"runtime"
 	"strings"
 
 	"github.com/okteto/okteto/cmd/utils"
@@ -57,6 +58,30 @@ type HybridExecCtx struct {
 	Name, Namespace string
 	Client          kubernetes.Interface
 	RunOktetoExec   bool
+}
+
+// GetCommandToExec for non-windows GOOS
+func (he *hybridExecutor) GetCommandToExec(ctx context.Context, cmd []string) (*exec.Cmd, error) {
+	var c *exec.Cmd
+	if runtime.GOOS != "windows" {
+		c = exec.Command(cmd[0], cmd[1:]...)
+	} else {
+		binary, err := expandExecutableInCurrentDirectory(cmd[0], he.workdir)
+		if err != nil {
+			return nil, err
+		}
+		c = exec.Command(binary, cmd[1:]...)
+	}
+
+	c.Env = he.envs
+
+	c.Stdin = os.Stdin
+	c.Stdout = os.Stdout
+	c.Stderr = os.Stderr
+
+	c.Dir = he.workdir
+
+	return c, nil
 }
 
 func (he *hybridExecutor) RunCommand(cmd *exec.Cmd) error {
